check management hub platform support

+++ build/common/src/utils.js
@@ -892,8 +892,13 @@ class Utils {
         return this.shell(arg);
     }
     purgeManagementHub() {
-        const arg = `curl -sSL https://raw.githubusercontent.com/open-horizon/devops/master/mgmt-hub/deploy-mgmt-hub.sh --output deploy-mgmt-hub.sh && chmod +x deploy-mgmt-hub.sh && sudo ./deploy-mgmt-hub.sh -PS && sudo rm -rf /tmp/horizon-all-in-1`;
-        return this.shell(arg);
+        if (os_1.default.arch() == 'x64' || process.platform == 'darwin') {
+            const arg = `curl -sSL https://raw.githubusercontent.com/open-horizon/devops/master/mgmt-hub/deploy-mgmt-hub.sh --output deploy-mgmt-hub.sh && chmod +x deploy-mgmt-hub.sh && sudo ./deploy-mgmt-hub.sh -PS && sudo rm -rf /tmp/horizon-all-in-1`;
+            return this.shell(arg);
+        }
+        else {
+            return (0, rxjs_1.of)();
+        }
     }
     cleanUp() {
         console.log('cleaning up', (0, fs_1.existsSync)(`${this.etcDefault}/horizon`), (0, fs_1.existsSync)(this.etcHorizon), this.etcHorizon);
